backend/requirements.txt
streamlit==1.31.1
groq==0.4.2
python-dotenv==1.0.0
mysql-connector-python==8.3.0
SQLAlchemy==2.0.27
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
boto3==1.34.34
psycopg2-binary==2.9.9
spacy==3.7.2
nltk==3.8.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pydantic==2.6.1
python-magic==0.4.27
aiofiles==23.2.1
python-magic-bin==0.4.14 